change types module to conversion; move z3 conversion type to conversion module
